feat(site): add Quick start terminal card to landing
Adds a "Quick start" section at the top of the homepage body (above the
existing docs CTA) showing `bun create mochi@latest` in a small
terminal-style card with traffic-light dots, sage `$` prompt, blinking
caret, and a copy-to-clipboard button that toggles to a check.

Also reverts the equivalent change accidentally applied to packages/demos
last commit — that's a separate site.
