+++ packages/demos/src/Landing.svelte
@@ -5,29 +5,7 @@
   import LayoutDashboard from '@lucide/svelte/icons/layout-dashboard';
   import ListChecks from '@lucide/svelte/icons/list-checks';
   import ArrowRight from '@lucide/svelte/icons/arrow-right';
-  import Copy from '@lucide/svelte/icons/copy';
-  import Check from '@lucide/svelte/icons/check';
   import type { Component } from 'svelte';
-
-  const QUICK_START_CMD = 'bun create mochi@latest';
-
-  let copied = $state(false);
-  let copyTimer: ReturnType<typeof setTimeout> | null = null;
-
-  function copyCommand() {
-    if (!navigator.clipboard) {
-      return;
-    }
-    navigator.clipboard.writeText(QUICK_START_CMD).then(() => {
-      copied = true;
-      if (copyTimer) {
-        clearTimeout(copyTimer);
-      }
-      copyTimer = setTimeout(() => {
-        copied = false;
-      }, 1600);
-    });
-  }
 
   type Demo = {
     href: string;
@@ -76,37 +54,6 @@
     <h1>Mochi Demos</h1>
     <p class="lede">A small collection of real apps built on the Mochi framework — server-rendered Svelte with islands of hydration.</p>
   </header>
-
-  <section class="quickstart" aria-labelledby="quickstart-title">
-    <header class="quickstart-head">
-      <h2 id="quickstart-title">Quick start</h2>
-      <p class="quickstart-hint">Scaffold a new Mochi project in seconds. Requires <a href="https://bun.sh" target="_blank" rel="noopener noreferrer">Bun</a>.</p>
-    </header>
-    <div class="terminal" role="figure" aria-label="Terminal command">
-      <div class="terminal-bar" aria-hidden="true">
-        <span class="terminal-dot terminal-dot-red"></span>
-        <span class="terminal-dot terminal-dot-amber"></span>
-        <span class="terminal-dot terminal-dot-green"></span>
-        <span class="terminal-title">~ &nbsp;—&nbsp; zsh</span>
-      </div>
-      <div class="terminal-body">
-        <div class="terminal-line">
-          <span class="terminal-prompt" aria-hidden="true">$</span>
-          <span class="terminal-cmd">{QUICK_START_CMD}</span>
-          <span class="terminal-caret" aria-hidden="true"></span>
-        </div>
-        <button class="terminal-copy" type="button" onclick={copyCommand} aria-label={copied ? 'Copied to clipboard' : 'Copy command'} title={copied ? 'Copied' : 'Copy command'}>
-          {#if copied}
-            <Check size={13} strokeWidth={2.2} />
-            <span class="terminal-copy-label">Copied</span>
-          {:else}
-            <Copy size={13} strokeWidth={1.8} />
-            <span class="terminal-copy-label">Copy</span>
-          {/if}
-        </button>
-      </div>
-    </div>
-  </section>
 
   <section class="grid" aria-label="Demo applications">
     {#each demos as demo (demo.href)}
@@ -183,199 +130,7 @@
   }
 
   .hero {
-    margin-bottom: clamp(1.75rem, 4vw, 2.5rem);
-  }
-
-  .quickstart {
-    margin-bottom: clamp(2.25rem, 5vw, 3.25rem);
-  }
-
-  .quickstart-head {
-    margin-bottom: 0.85rem;
-  }
-
-  .quickstart-head h2 {
-    font-family: var(--l-font-serif);
-    font-weight: 500;
-    font-variation-settings: 'opsz' 36;
-    font-size: 1.4rem;
-    line-height: 1.1;
-    letter-spacing: -0.015em;
-    color: var(--l-text);
-  }
-
-  .quickstart-hint {
-    margin-top: 0.35rem;
-    font-size: 0.92rem;
-    color: var(--l-text-muted);
-  }
-
-  .quickstart-hint a {
-    color: var(--l-accent);
-    text-decoration: none;
-    border-bottom: 1px solid transparent;
-    transition:
-      color 120ms ease,
-      border-color 120ms ease;
-  }
-
-  .quickstart-hint a:hover {
-    border-bottom-color: var(--l-accent);
-  }
-
-  .terminal {
-    background: #1f2a24;
-    border: 1px solid #0f1612;
-    border-radius: var(--l-radius-lg);
-    box-shadow:
-      0 14px 36px rgba(15, 22, 18, 0.18),
-      0 2px 6px rgba(15, 22, 18, 0.08),
-      0 0 0 1px rgba(255, 255, 255, 0.03) inset;
-    overflow: hidden;
-  }
-
-  .terminal-bar {
-    display: flex;
-    align-items: center;
-    gap: 0.4rem;
-    padding: 0.55rem 0.85rem;
-    background: rgba(0, 0, 0, 0.22);
-    border-bottom: 1px solid rgba(255, 255, 255, 0.04);
-  }
-
-  .terminal-dot {
-    width: 0.72rem;
-    height: 0.72rem;
-    border-radius: 999px;
-    box-shadow: 0 0 0 1px rgba(0, 0, 0, 0.25) inset;
-  }
-
-  .terminal-dot-red {
-    background: #ff5f57;
-  }
-
-  .terminal-dot-amber {
-    background: #febc2e;
-  }
-
-  .terminal-dot-green {
-    background: #28c840;
-  }
-
-  .terminal-title {
-    margin-left: 0.6rem;
-    font-family: var(--l-font-mono);
-    font-size: 0.72rem;
-    color: rgba(232, 230, 221, 0.45);
-    letter-spacing: 0.03em;
-  }
-
-  .terminal-body {
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-    padding: 1.05rem 1.15rem 1.05rem 1.25rem;
-    font-family: var(--l-font-mono);
-    font-size: 0.98rem;
-    line-height: 1.4;
-    color: #e8e6dd;
-  }
-
-  .terminal-line {
-    flex: 1;
-    min-width: 0;
-    display: inline-flex;
-    align-items: center;
-    gap: 0.55rem;
-    overflow-x: auto;
-    white-space: nowrap;
-  }
-
-  .terminal-prompt {
-    color: #8ab79a;
-    font-weight: 600;
-    user-select: none;
-  }
-
-  .terminal-cmd {
-    color: #e8e6dd;
-    user-select: text;
-  }
-
-  .terminal-caret {
-    display: inline-block;
-    width: 0.5rem;
-    height: 1.05em;
-    background: #8ab79a;
-    border-radius: 1px;
-    animation: terminal-caret-blink 1.1s steps(2, end) infinite;
-    transform: translateY(0.12em);
-    opacity: 0.85;
-  }
-
-  @keyframes terminal-caret-blink {
-    0%,
-    50% {
-      opacity: 0.85;
-    }
-    51%,
-    100% {
-      opacity: 0;
-    }
-  }
-
-  .terminal-copy {
-    flex: 0 0 auto;
-    display: inline-flex;
-    align-items: center;
-    gap: 0.35rem;
-    background: rgba(232, 230, 221, 0.08);
-    border: 1px solid rgba(232, 230, 221, 0.14);
-    color: #d4d2c7;
-    font-family: var(--l-font-sans);
-    font-size: 0.75rem;
-    font-weight: 500;
-    padding: 0.4rem 0.7rem;
-    border-radius: var(--l-radius-md);
-    cursor: pointer;
-    transition:
-      background 120ms ease,
-      border-color 120ms ease,
-      color 120ms ease,
-      transform 120ms ease;
-    letter-spacing: 0.02em;
-  }
-
-  .terminal-copy:hover {
-    background: rgba(232, 230, 221, 0.14);
-    border-color: rgba(232, 230, 221, 0.24);
-    color: #ffffff;
-  }
-
-  .terminal-copy:active {
-    transform: translateY(1px);
-  }
-
-  .terminal-copy:focus-visible {
-    outline: 2px solid #8ab79a;
-    outline-offset: 2px;
-  }
-
-  .terminal-copy-label {
-    line-height: 1;
-  }
-
-  @media (max-width: 540px) {
-    .terminal-body {
-      padding: 0.9rem 0.95rem;
-      font-size: 0.85rem;
-    }
-    .terminal-copy-label {
-      display: none;
-    }
-    .terminal-copy {
-      padding: 0.4rem 0.5rem;
-    }
+    margin-bottom: clamp(2.5rem, 6vw, 4rem);
   }
 
   .hero h1 {
